LOGGED-IN and LOGGED-IN-ROLE
Fixed some security issues with the teacher and student pages.

Added a logged-in and a role checker to auth and portal.py... which makes sure that 1)the user is logged in and 2) the user that is logged in is a teacher, in order to view any of the teacher pages.

added a go-back and home button to the base html.... teachers logged in will be taken to the teacher home page and students logged in will be taken to the student home page