FIXED CSS AND CHANGED ABORT MESSAGE
ended up having to completely overhaul the html and css to get the curses and sessions to display correctly, but i think what we have now is MUCH more developer friendly,
and it just makes sence. instaed of a series of divs and worrying about positioning, i'm returning all the course and session information in unordered lists. the code is much cleaner
and easier to read.

the problem anderw found with the abort was fixed, it now returns a proper error and abort message.